Refactor DAO tests and improve file handling logic.
Updated test cases to use inline object creation for `Sinistro` and `Apolice`. Enhanced file deletion logic in `TesteDAO` to handle null and non-existing directories gracefully, improving code robustness and readability.

+++ src/br/edu/cs/poo/ac/seguro/testes/TesteDAO.java
@@ -5,8 +5,7 @@ import java.io.File;
 import org.junit.jupiter.api.BeforeEach;
 
 import br.edu.cesarschool.next.oo.persistenciaobjetos.CadastroObjetos;
-import br.edu.cs.poo.ac.seguro.daos.SeguradoEmpresaDAO;
-import br.edu.cs.poo.ac.seguro.entidades.SeguradoEmpresa;
+
 
 public abstract class TesteDAO {
     protected CadastroObjetos cadastro;
@@ -21,9 +20,14 @@ public abstract class TesteDAO {
     public void setUp() {
         String sep = File.separator;
         File dir = new File("." + sep + getClasse().getSimpleName());
-        File[] files = dir.listFiles();
-        for (File file : files) {
-            file.delete();
+
+        if (dir.exists()) {
+            File[] files = dir.listFiles();
+            if (files != null) {
+                for (File file : files) {
+                    file.delete();
+                }
+            }
         }
     }
 }
